requirements.txt
pillow==10.4.0
six==1.16.0
wxPython==4.2.2
